# server/swagger.yaml
openapi: 3.0.3
info:
  title: Rocko Backend API Docs
  version: 1.0.1
servers:
  - url: https://backend.dev.rocko.cloud
  - url: http://localhost:5000
  - url: https://backend.prod.rocko.cloud
...
paths:
  /loans:
    get:
      tags:
        - loan
      summary: ""
      description: ""
      parameters:
        - name: user
          in: query
          description: ""
          required: false
          explode: true
          schema:
            type : integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
        '400':
          description: Invalid status value
      security:
        - bearerAuth: [] 
  /add:
    post:
      tags:
        - loan
      summary: ''
      description: ''
      requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddLoan'
            application/xml:
              schema:
                $ref: '#/components/schemas/AddLoan'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/AddLoan'
      responses:
        '200':
          description: Successful operation
      security:
        - bearerAuth: [] 
  /update:
    post:
      tags:
        - loan
      requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLoan'
            application/xml:
              schema:
                $ref: '#/components/schemas/UpdateLoan'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/UpdateLoan'
      responses:
        '200':
           description: Successful operation
      security:
        - bearerAuth: [] 
...
  /comp/address:
    post:
      tags:
        - compliance
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                    example: "0x616074f143306b4CeFe272E546f73044e85C6d6d"
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad Reques
        '401':
          description: Unauthorized
      security:
        - bearerAuth: [] 
  /comp/transaction:
    post:
      tags:
        - compliance
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_hash:
                    type: string
                    example: "0xb32cc3ef574e79257d2a87a0a04c53efc798c3f9c762c4924b6645175ce093f5"
                  destination_address:
                    type: string
                    example: "0xX"
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad Reques
        '401':
          description: Unauthorized
      security:
        - bearerAuth: [] 
  /comp/platform-status:
    get:
      tags:
        - compliance
      summary: ""
      description: ""
      responses:
        '200':
          description: ""
          content:
              application/json:
                schema:
                  type: object
                  properties:
                    loan_booking_blocked:
                      type: number
                    transactions_blocked:
                      type: number
        '400':
          description: Bad Reques
        '401':
          description: Unauthorized
...
components:
  schemas:
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        user_id:
          type: integer
          format: int64
          example: 21
        transaction_hash:
          type: string
          example: "0x2f6360e8e282dbf361c3a5d125d9b4e3568505f7186c9334c3347dea4782fe7a"
        lending_protocol:
          type: string
        loan_active:
          type: integer
        loan_asset:
          type: string
        principal_balance: 
          type: number
        outstanding_balance:
          type: number
        collateral: 
          type: integer
        interest:
          type: number
        collateral_decimals:
          type: integer
        created_time:
          type: string
      xml:
        name: Loan
    AddLoan:
      type: object
      properties:
        user:
          type: integer
          format: int64
          example: 26
        transaction_hash:
          type: string
          example: "0x2f6360e8e282dbf361c3a5d125d9b4e3568505f7186c9334c3347dea4782fe7a"
        lending_protocol:
          type: string
          example: compound
        loan_active:
          type: boolean
          example: 1
        loan_asset:
          type: string
          example: ETH
        outstanding_balance:
          type: number
          example: 3000.000000
        collateral: 
          type: integer
          example: 182202000000000000
        exist:
          type: boolean
          example: 0
      xml:
        name: AddLoan
    UpdateLoan:
      type: object
      properties:
        updateType:
          type: string
          example: repay
        id:
          type: integer
          format: int64
          example: 26
        loan_active:
          type: boolean
          example: 1
        outstanding_balance:
          type: number
          example: 3000.000000
        collateral: 
          type: integer
          example: 182202000000000000
        interest: 
          type: number
          example: 0.000000
        transaction_hash:
          type: string
          example: "0x2f6360e8e282dbf361c3a5d125d9b4e3568505f7186c9334c3347dea4782fe7a"
      xml:
        name: UpdateLoan
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 